requirements.txt
streamlit
python-dotenv
google-generativeai==0.8.5
matplotlib
requests
